+++ NEWS.md
@@ -5,6 +5,7 @@
 
 ## rempsyc 0.1.0.9
 * `nice_assumptions`: now supports list objects, and does not print the interpretation message anymore since it is available from the documentation.
+* `nice_var`: now supports list objects.
 
 ## rempsyc 0.1.0.8
 * `nice_violin` and `nice_contrasts`: Now correctly handle missing values.